automate-the-boring-stuff/requirements.txt
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
lxml==4.4.2
mccabe==0.6.1
MouseInfo==0.1.2
numpy==1.17.4
openpyxl==3.0.2
Pillow==6.2.1
PyAutoGUI==0.9.48
pycodestyle==2.5.0
PyGetWindow==0.0.8
PyJWT==1.7.1
pylint==2.4.4
PyMsgBox==1.0.7
PyPDF2==1.26.0
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.25
python-docx==0.8.10
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
requests==2.22.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
twilio==6.35.0
urllib3==1.25.7
wrapt==1.11.2
